Add toastr services

+++ src/app/header/header.component.ts
@@ -10,13 +10,18 @@ export class HeaderComponent {
   constructor(
     private router: Router,
   ){ }
-   isLogin = sessionStorage.getItem("token") ? true : false
+
+  isLogin = sessionStorage.getItem("token") ? true : false
 
   logoutUser() {
-    if(confirm('are you sure you want to logout this user')){
-      sessionStorage.clear()
-      this.router.navigateByUrl('/login')
-      location.reload()
+    if (this.isLogin == true) {
+      if(confirm('are you sure you want to logout this user')){
+        sessionStorage.clear()
+        this.router.navigateByUrl('/login')
+        location.reload()
+      }
+    } else {
+      console.error('')
     }
   }
 }
